+++ psd-web/app/models/audit_activity/investigation/team_added.rb
@@ -1,6 +1,6 @@
 class AuditActivity::Investigation::TeamAdded < AuditActivity::Investigation::Base
-  def subtitle
-    "Team added by #{source&.show}, #{pretty_date_stamp}"
+  def subtitle(viewer)
+    "Team added by #{source&.show(viewer)}, #{pretty_date_stamp}"
   end
 
 private