+++ psd-web/app/models/investigation.rb
@@ -20,15 +20,13 @@ class Investigation < ApplicationRecord
   validates :type, presence: true # Prevent saving instances of Investigation; must use a subclass instead
 
   validates :description, presence: true, on: :update
-  # validates :owner_id, presence: { message: "Select case owner" }, on: :update
 
   validates :user_title, length: { maximum: 100 }
   validates :description, length: { maximum: 10_000 }
   validates :non_compliant_reason, length: { maximum: 10_000 }
   validates :hazard_description, length: { maximum: 10_000 }
 
-  after_update :create_audit_activity_for_owner,
-               :create_audit_activity_for_status,
+  after_update :create_audit_activity_for_status,
                :create_audit_activity_for_visibility,
                :create_audit_activity_for_summary
 
@@ -47,6 +45,10 @@ class Investigation < ApplicationRecord
 
   has_many :corrective_actions, dependent: :destroy
   has_many :correspondences, dependent: :destroy
+  has_many :emails, dependent: :destroy, class_name: "Correspondence::Email"
+  has_many :phone_calls, dependent: :destroy, class_name: "Correspondence::PhoneCall"
+  has_many :meetings, dependent: :destroy, class_name: "Correspondence::Meeting"
+
   has_many :tests, dependent: :destroy
   has_many :test_results, class_name: "Test::Result", dependent: :destroy
   has_many :alerts, dependent: :destroy
@@ -114,8 +116,7 @@ class Investigation < ApplicationRecord
 
   def past_owners
     activities = AuditActivity::Investigation::UpdateOwner.where(investigation_id: id)
-    user_id_list = activities.map(&:owner_id)
-    User.where(id: user_id_list)
+    activities.map(&:owner)
   end
 
   def important_owner_teams
@@ -126,12 +127,6 @@ class Investigation < ApplicationRecord
     end
     teams << owner if owner.is_a? Team
     teams
-  end
-
-  def past_teams
-    activities = AuditActivity::Investigation::UpdateOwner.where(investigation_id: id)
-    team_id_list = activities.map(&:owner_id)
-    Team.where(id: team_id_list)
   end
 
   def enquiry?
@@ -186,14 +181,6 @@ private
   def create_audit_activity_for_visibility
     if saved_changes.key?(:is_private) || visibility_rationale.present?
       AuditActivity::Investigation::UpdateVisibility.from(self)
-    end
-  end
-
-  def create_audit_activity_for_owner
-    # TODO: User.current check is here to avoid triggering activity and emails from migrations
-    # Can be safely removed once the migration PopulateAssigneeAndDescription has run
-    if @owner_changed && User.current
-      AuditActivity::Investigation::UpdateOwner.from(self)
     end
   end