+++ psd-web/spec/features/record_email_activity_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
 
   let(:name) { "Test name" }
   let(:email) { Faker::Internet.safe_email }
-  let(:date) { Time.zone.today }
+  let(:date) { Date.parse("2020-02-01") }
 
   let(:file) { Rails.root.join("test/fixtures/files/attachment_filename.txt") }
   let(:summary) { "Test summary" }
@@ -64,6 +64,7 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     click_button "Continue"
 
     expect_to_be_on_record_email_details_page
+    fill_in "Summary", with: "Test summary"
 
     within_fieldset "Email content" do
       expect(page).to have_css("a", text: File.basename(file))
@@ -80,7 +81,15 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
 
     # Consumer info is not hidden from case owner
-    expect_case_activity_page_to_show_entered_information(name: name, email: email, date: date, file: file)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, email: email, date: date, file: file)
+
+    click_link "View email"
+
+    expect_to_be_on_email_page(case_id: investigation.pretty_id)
+    expect(page).to have_h1("Test summary")
+    expect(page).to have_summary_item(key: "Date of email", value: "1 February 2020")
+    expect(page).to have_summary_item(key: "From", value: "#{name} (#{email})")
+    expect(page).to have_summary_item(key: "Email", value: "attachment_filename.txt (0 Bytes)")
 
     # Test that another user in a different organisation cannot see consumer info
     sign_out
@@ -100,7 +109,7 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     visit "/cases/#{investigation.pretty_id}/activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, email: email, date: date, file: file)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, email: email, date: date, file: file)
   end
 
   scenario "with non-consumer contact details and summary and subject and body" do
@@ -126,7 +135,7 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     click_on "Activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, email: email, date: date, summary: summary, subject: email_subject, body: body)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, email: email, date: date, summary: summary, subject: email_subject, body: body)
 
     # Test that another user in a different organisation can see all info
     sign_out
@@ -136,7 +145,7 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     visit "/cases/#{investigation.pretty_id}/activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, email: email, date: date, summary: summary, subject: email_subject, body: body)
+    expect_case_activity_page_to_show_entered_information(user_name: "#{user.name} (#{user.team.name})", name: name, email: email, date: date, summary: summary, subject: email_subject, body: body)
   end
 
   def fill_in_record_email_form(name:, email:, consumer:, date:)
@@ -189,14 +198,14 @@ RSpec.feature "Adding a record email activity to a case", :with_stubbed_elastics
     expect(find_field("Year").value).to eq date.year.to_s
   end
 
-  def expect_case_activity_page_to_show_entered_information(name:, email:, date:, file: nil, summary: nil, subject: nil, body: nil)
-    item = page.find("p", text: "Email recorded by #{user.name} (#{user.team.name})").find(:xpath, "..")
+  def expect_case_activity_page_to_show_entered_information(user_name:, name:, email:, date:, file: nil, summary: nil, subject: nil, body: nil)
+    item = page.find("p", text: "Email recorded by #{user_name}").find(:xpath, "..")
     expect(item).to have_text("From: #{name} (#{email})")
     expect(item).to have_text("Date sent: #{date.strftime('%d/%m/%Y')}")
 
     if file
       expect(item).to have_text("Email: #{File.basename(file)}")
-      expect(item).to have_link("View email file")
+      expect(item).to have_link("View email")
     else
       expect(item).to have_text(summary)
       expect(item).to have_text(subject)