+++ psd-web/spec/features/record_phone_call_activity_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
 
   let(:name) { "Test name" }
   let(:phone) { "07000 000000" }
-  let(:date) { Time.zone.today }
+  let(:date) { Date.parse("2020-05-05") }
   let(:file) { Rails.root.join("test/fixtures/files/attachment_filename.txt") }
   let(:summary) { "Test summary" }
   let(:notes) { "Test notes" }
@@ -78,7 +78,15 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
 
     # Consumer info is not hidden from case owner
-    expect_case_activity_page_to_show_entered_information(name: name, phone: phone, date: date, file: file)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, phone: phone, date: date, file: file)
+
+    click_link "View phone call"
+
+    expect_to_be_on_phone_call_page(case_id: investigation.pretty_id)
+
+    expect(page).to have_summary_item(key: "Date of call", value: "5 May 2020")
+    expect(page).to have_summary_item(key: "Call with", value: "#{name} (#{phone})")
+    expect(page).to have_summary_item(key: "Transcript", value: "attachment_filename.txt (0 Bytes)")
 
     # Test that another user in a different organisation cannot see consumer info
     sign_out
@@ -98,7 +106,7 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
     visit "/cases/#{investigation.pretty_id}/activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, phone: phone, date: date, file: file)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, phone: phone, date: date, file: file)
   end
 
   scenario "with non-consumer contact details and summary and notes" do
@@ -124,7 +132,7 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
     click_on "Activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, phone: phone, date: date, summary: summary, notes: notes)
+    expect_case_activity_page_to_show_entered_information(user_name: user.name, name: name, phone: phone, date: date, summary: summary, notes: notes)
 
     # Test that another user in a different organisation can see all info
     sign_out
@@ -134,7 +142,7 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
     visit "/cases/#{investigation.pretty_id}/activity"
 
     expect_to_be_on_case_activity_page(case_id: investigation.pretty_id)
-    expect_case_activity_page_to_show_entered_information(name: name, phone: phone, date: date, summary: summary, notes: notes)
+    expect_case_activity_page_to_show_entered_information(user_name: "#{user.name} (#{user.team.name})", name: name, phone: phone, date: date, summary: summary, notes: notes)
   end
 
   def fill_in_record_phone_call_form(name:, phone:, consumer:, date:)
@@ -181,14 +189,14 @@ RSpec.feature "Adding a record phone call activity to a case", :with_stubbed_ela
     expect(find_field("Year").value).to eq date.year.to_s
   end
 
-  def expect_case_activity_page_to_show_entered_information(name:, phone:, date:, file: nil, summary: nil, notes: nil)
-    item = page.find("p", text: "Phone call by #{user.name} (#{user.team.name})").find(:xpath, "..")
+  def expect_case_activity_page_to_show_entered_information(user_name:, name:, phone:, date:, file: nil, summary: nil, notes: nil)
+    item = page.find("p", text: "Phone call by #{user_name}").find(:xpath, "..")
     expect(item).to have_text("Call with: #{name} (#{phone})")
     expect(item).to have_text("Date: #{date.strftime('%d/%m/%Y')}")
 
     if file
       expect(item).to have_text("Attached: #{File.basename(file)}")
-      expect(item).to have_link("View attachment")
+      expect(item).to have_link("View phone call")
     else
       expect(item).to have_text(summary)
       expect(item).to have_text(notes)