+++ psd-web/spec/models/audit_activity/investigation/update_owner_spec.rb
@@ -1,21 +1,39 @@
 require "rails_helper"
 
-RSpec.describe AuditActivity::Investigation::UpdateOwner, :with_stubbed_mailer, :with_stubbed_elasticsearch do
-  subject(:audit_activity) { described_class.from(investigation) }
+RSpec.describe AuditActivity::Investigation::UpdateOwner, :with_stubbed_elasticsearch, :with_stubbed_mailer do
+  subject(:activity) { described_class.create(investigation: investigation, metadata: described_class.build_metadata(owner, rationale)) }
 
-  let(:user) { create(:user).decorate }
-  let(:investigation) { create(:enquiry, owner: user, owner_rationale: "Test owner") }
+  let(:investigation) { create(:allegation, owner: owner) }
+  let(:owner) { create(:team) }
+  let(:rationale) { "Test rationale" }
+  let(:user) { create(:user) }
 
-  before { User.current = user }
+  describe ".build_metadata" do
+    let(:result) { described_class.build_metadata(owner, rationale) }
 
-  describe ".from" do
-    it "creates an audit activity with information for email notification", :aggregate_failures do
-      expect(audit_activity).to have_attributes(
-        title: "Case owner changed to #{user.display_name}",
-        body: investigation.owner_rationale,
-        email_subject_text: "Case owner changed for enquiry"
-      )
-      expect(audit_activity.email_update_text(user)).to start_with("Case owner changed on enquiry to #{user.name} by #{user.name}")
+    it "returns a Hash of the arguments" do
+      expect(result).to eq({
+        owner_id: owner.id,
+        rationale: rationale
+      })
+    end
+  end
+
+  describe "#title" do
+    it "returns a generated String" do
+      expect(activity.title(user)).to eq("Case owner changed to #{owner.name}")
+    end
+  end
+
+  describe "#body" do
+    it "returns the rationale" do
+      expect(activity.body).to eq("Test rationale")
+    end
+  end
+
+  describe "#owner" do
+    it "returns the owner" do
+      expect(activity.owner).to eq(owner)
     end
   end
 end