+++ psd-web/spec/models/user_spec.rb
@@ -163,7 +163,7 @@ RSpec.describe User do
     it "includes associations needed for display_name" do
       owners = described_class.get_owners.to_a # to_a forces the query execution and load immediately
       expect(lambda {
-        owners.map { |owner| owner.decorate.display_name }
+        owners.map { |owner| owner.decorate.display_name(viewer: active_user) }
       }).to not_talk_to_db
     end