+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2020-06-12
+- Users of all teams added to a case can now view complainant contact details on the case activity page.
+
+## 2020-06-05
+– Added the ability to view the details of correspondence added to the case on their own page.
+
 ## 2020-05-26
 - Update service guidance to give details of revised permission model on cases.